+++ src/ESP32_PWM.h
@@ -21,11 +21,12 @@
   Therefore, their executions are not blocked by bad-behaving functions / tasks.
   This important feature is absolutely necessary for mission-critical tasks.
 
-  Version: 1.0.0
+  Version: 1.0.1
 
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0   K Hoang      20/09/2021 Initial coding for ESP32, ESP32_S2, ESP32_C3 boards with ESP32 core v2.0.0+
+  1.0.1   K Hoang      21/09/2021 Fix bug. Ading PWM end-of-duty-cycle callback function. Improve examples
 *****************************************************************************************************************************/
 
 #pragma once
@@ -52,7 +53,7 @@
 #endif
 
 #ifndef ESP32_PWM_VERSION
-  #define ESP32_PWM_VERSION       "ESP32_PWM v1.0.0"
+  #define ESP32_PWM_VERSION       "ESP32_PWM v1.0.1"
 #endif
 
 #ifndef TIMER_INTERRUPT_DEBUG
@@ -175,7 +176,7 @@ typedef ESP32TimerInterrupt ESP32Timer;
 //esp_err_t timer_group_intr_enable(timer_group_t group_num, timer_intr_t intr_mask);
 //esp_err_t timer_group_intr_disable(timer_group_t group_num, timer_intr_t intr_mask);
 
-
+#define INVALID_ESP32_PIN         255
 
 typedef bool (*esp32_timer_callback)  (void *);